fix: InventorySystem 모듈에서 ItemSystem 모듈 의존성 제거
