+++ Source/GGFInventorySystem/Private/GGFInventoryItemConfig.cpp
@@ -1,0 +1,4 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "GGFInventoryItemConfig.h"
